# account_manager/openapi/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: vasco.m.campos@inesctec.pt
    name: Vasco Campos
  description: 'User Account Manager Service OpenAPI definition. This service has
    the following functions: register, login, change user credentials and update it''s
    profile settings. Find out more: [User Account Manager Service documentation](https://gitlab.inesctec.pt/cpes/european-projects/interconnect/hems/hems-documentation/-/blob/master/Microservices/User-Account-Manager-Service.adoc)'
  title: User Account Manager Service
  version: 1.2.0
servers:
- url: https://interconnect-dev.inesctec.pt/api/account
security:
- {}
- Bearer: []
...
paths:
...
  /login:
    description: Login
    post:
      operationId: login_post
      parameters:
      - explode: false
        in: header
        name: X-Correlation-ID
        required: true
        schema:
          $ref: '#/components/schemas/CorrelationId'
        style: simple
      - explode: false
        in: header
        name: expo-token
        required: false
        schema:
          $ref: '#/components/schemas/ExpoToken'
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        "200":
          description: Successfully authenticated. The token is in the Authorization
            header. You need to include this header is subsequent requests.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
            Authorization:
              $ref: '#/components/headers/Authorization'
        "400":
          content:
            application/json:
              examples:
                Bad format:
                  value:
                    error: Invalid specification format
                Timestamp error:
                  value:
                    error: one or more timestamps in the request are invalid
                Missing X-Correlation-ID header:
                  value:
                    error: missing X-Correlation-ID header
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "401":
          content:
            application/json:
              examples:
                Unauthenticated:
                  value:
                    error: error in authorization token
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthenticated or invalid credentials
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
      summary: Login
      tags:
      - Session management
      x-openapi-router-controller: account_manager.controllers.session_management_controller
  /logout:
    description: Logout
    post:
      operationId: logout_post
      parameters:
      - explode: false
        in: header
        name: X-Correlation-ID
        required: true
        schema:
          $ref: '#/components/schemas/CorrelationId'
        style: simple
      - explode: false
        in: header
        name: Authorization
        required: true
        schema:
          $ref: '#/components/schemas/Authorization'
        style: simple
      responses:
        "200":
          description: User logout
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "400":
          content:
            application/json:
              examples:
                Bad format:
                  value:
                    error: Invalid specification format
                Timestamp error:
                  value:
                    error: one or more timestamps in the request are invalid
                Missing X-Correlation-ID header:
                  value:
                    error: missing X-Correlation-ID header
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "401":
          content:
            application/json:
              examples:
                Unauthenticated:
                  value:
                    error: error in authorization token
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthenticated or invalid credentials
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
      summary: Logout
      tags:
      - Session management
      x-openapi-router-controller: account_manager.controllers.session_management_controller
...
  /refresh-token:
    description: Refresh authentication token (needed if the token is about to expire)
    post:
      operationId: refresh_token_post
      parameters:
      - explode: false
        in: header
        name: X-Correlation-ID
        required: true
        schema:
          $ref: '#/components/schemas/CorrelationId'
        style: simple
      - explode: false
        in: header
        name: Authorization
        required: true
        schema:
          $ref: '#/components/schemas/Authorization'
        style: simple
      responses:
        "200":
          description: Successfully generated a new token. The new token is in the
            Authorization header.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
            Authorization:
              $ref: '#/components/headers/Authorization'
        "400":
          content:
            application/json:
              examples:
                Bad format:
                  value:
                    error: Invalid specification format
                Timestamp error:
                  value:
                    error: one or more timestamps in the request are invalid
                Missing X-Correlation-ID header:
                  value:
                    error: missing X-Correlation-ID header
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        "401":
          content:
            application/json:
              examples:
                Unauthenticated:
                  value:
                    error: error in authorization token
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthenticated or invalid credentials
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
      summary: Refresh authentication token (needed if the token is about to expire)
      tags:
      - Session management
      x-openapi-router-controller: account_manager.controllers.session_management_controller
...
  /register-google:
    description: Register/Login user with google account
    post:
      operationId: register_google_post
      parameters:
      - explode: false
        in: header
        name: X-Correlation-ID
        required: true
        schema:
          $ref: '#/components/schemas/CorrelationId'
        style: simple
      - explode: false
        in: header
        name: Authorization
        required: true
        schema:
          $ref: '#/components/schemas/Authorization'
        style: simple
      - explode: false
        in: header
        name: expo-token
        required: false
        schema:
          $ref: '#/components/schemas/ExpoToken'
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleUserRegisterRequest'
        required: true
      responses:
        "200":
          description: Successfully authenticated. The token is in the Authorization
            header. You need to include this header in subsequent requests.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
            Authorization:
              $ref: '#/components/headers/Authorization'
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: Account created and google login is also made. The token is
            in the Authorization header.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
            Authorization:
              $ref: '#/components/headers/Authorization'
        "400":
          content:
            application/json:
              examples:
                Bad format:
                  value:
                    error: Invalid specification format
                Timestamp error:
                  value:
                    error: one or more timestamps in the request are invalid
                Missing X-Correlation-ID header:
                  value:
                    error: missing X-Correlation-ID header
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
      summary: Register/Login user with google account
      tags:
      - Session management
      x-openapi-router-controller: account_manager.controllers.session_management_controller
...
components:
  headers:
    CorrelationId:
      explode: false
      required: true
      schema:
        $ref: '#/components/schemas/CorrelationId'
      style: simple
    Authorization:
      explode: false
      required: true
      schema:
        $ref: '#/components/schemas/Authorization'
      style: simple
...
  schemas:
    CorrelationId:
      format: uuid
      type: string
...
    Authorization:
      format: byte
      type: string
...
    Country:
      enum:
      - PT
      example: PT
      title: Country
      type: string
...
    GoogleUserRegisterRequest:
      description: Request body for a google user
      example:
        last_name: Ana
        first_name: Ana
        email: myemail@mycompany.com
      properties:
        first_name:
          example: Ana
          maxLength: 15
          minLength: 1
          title: first_name
          type: string
        last_name:
          example: Ana
          maxLength: 15
          minLength: 1
          title: first_name
          type: string
        email:
          example: myemail@mycompany.com
          format: email
          maxLength: 50
          title: email
          type: string
      required:
      - email
      - first_name
      type: object
    LoginRequest:
      description: Model containing the login request
      example:
        password: Ja7ha6dUHs7knm
        email: myemail@mycompany.com
      properties:
        email:
          example: myemail@mycompany.com
          format: email
          maxLength: 50
          title: email
          type: string
        password:
          example: Ja7ha6dUHs7knm
          format: password
          maxLength: 30
          minLength: 6
          type: string
      required:
      - email
      - password
      type: object
...
    ScheduleType:
      enum:
      - economic
      - ecologic
      example: economic
      title: ScheduleType
      type: string
    TarifType:
      enum:
      - simple
      - bi-hourly
      - tri-hourly
      example: simple
      title: TarifType
      type: string
    ContractedPower:
      enum:
      - 1.15 kVA
      - 2.3 kVA
      - 3.45 kVA
      - 4.6 kVA
      - 5.75 kVA
      - 6.9 kVA
      - 10.35 kVA
      - 13.8 kVA
      - 17.25 kVA
      - 20.7 kVA
      - 27.6 kVA
      - 34.5 kVA
      - 41.4 kVA
      - 27.6 kVA
      - 34.5 kVA
      - 41.4 kVA
      example: 6.9 kVA
      title: ContractedPower
      type: string
...
    Permission:
      enum:
      - None
      - Minimal
      - Full
      example: Full
      title: Permission
      type: string
...
    ExpoToken:
      example: I22V7MDFJx
      title: expo_token
      type: string
...
    User:
      description: Settings object
      example:
        country: PT
        is_active: true
        schedule_type: economic
        not_disturb:
          sunday:
          - end_timestamp: 2000-01-23T04:56:07.000+00:00
            start_timestamp: 2000-01-23T04:56:07.000+00:00
          - end_timestamp: 2000-01-23T04:56:07.000+00:00
            start_timestamp: 2000-01-23T04:56:07.000+00:00
          saturday:
          - end_timestamp: 2000-01-23T04:56:07.000+00:00
            start_timestamp: 2000-01-23T04:56:07.000+00:00
          - end_timestamp: 2000-01-23T04:56:07.000+00:00
            start_timestamp: 2000-01-23T04:56:07.000+00:00
          tuesday:
          - end_timestamp: 2000-01-23T04:56:07.000+00:00
            start_timestamp: 2000-01-23T04:56:07.000+00:00
          - end_timestamp: 2000-01-23T04:56:07.000+00:00
            start_timestamp: 2000-01-23T04:56:07.000+00:00
          wednesday:
          - end_timestamp: 2000-01-23T04:56:07.000+00:00
            start_timestamp: 2000-01-23T04:56:07.000+00:00
          - end_timestamp: 2000-01-23T04:56:07.000+00:00
            start_timestamp: 2000-01-23T04:56:07.000+00:00
          thursday:
          - end_timestamp: 2000-01-23T04:56:07.000+00:00
            start_timestamp: 2000-01-23T04:56:07.000+00:00
          - end_timestamp: 2000-01-23T04:56:07.000+00:00
            start_timestamp: 2000-01-23T04:56:07.000+00:00
          friday:
          - end_timestamp: 2000-01-23T04:56:07.000+00:00
            start_timestamp: 2000-01-23T04:56:07.000+00:00
          - end_timestamp: 2000-01-23T04:56:07.000+00:00
            start_timestamp: 2000-01-23T04:56:07.000+00:00
          monday:
          - end_timestamp: 2000-01-23T04:56:07.000+00:00
            start_timestamp: 2000-01-23T04:56:07.000+00:00
          - end_timestamp: 2000-01-23T04:56:07.000+00:00
            start_timestamp: 2000-01-23T04:56:07.000+00:00
        is_google_account: true
        last_name: Ana
        wp_token: token
        modified_timestamp: 2000-01-23T04:56:07.000+00:00
        contracted_power: 6.9 kVA
        expo_token: I22V7MDFJx
        user_id: userId1
        api_key: api_key
        permissions: Full
        global_optimizer: true
        postal_code: postal_code
        first_name: Ana
        meter_id: meter_id
        email: myemail@mycompany.com
        tarif_type: simple
      properties:
        user_id:
          example: userId1
          maxLength: 10
          minLength: 10
          title: user_id
          type: string
        first_name:
          example: Ana
          maxLength: 15
          minLength: 1
          title: first_name
          type: string
        last_name:
          example: Ana
          maxLength: 15
          minLength: 1
          title: first_name
          type: string
        email:
          example: myemail@mycompany.com
          format: email
          maxLength: 50
          title: email
          type: string
        is_active:
          title: is_active
          type: boolean
        schedule_type:
          $ref: '#/components/schemas/ScheduleType'
        meter_id:
          title: MeterId
          type: string
        api_key:
          title: api_key
          type: string
        expo_token:
          example: I22V7MDFJx
          title: expo_token
          type: string
        wp_token:
          example: token
          title: wp_token
          type: string
        country:
          $ref: '#/components/schemas/Country'
        postal_code:
          pattern: ^[0-9]{4}-[0-9]{3}$
          title: postal_code
          type: string
        tarif_type:
          $ref: '#/components/schemas/TarifType'
        contracted_power:
          $ref: '#/components/schemas/ContractedPower'
        not_disturb:
          $ref: '#/components/schemas/NotDisturb'
        global_optimizer:
          title: global_optimizer
          type: boolean
        permissions:
          $ref: '#/components/schemas/Permission'
        is_google_account:
          example: true
          title: is_google_account
          type: boolean
        modified_timestamp:
          format: date-time
          title: start_timestamp
          type: string
      required:
      - email
      - first_name
      - is_active
      - is_google_account
      - modified_timestamp
      - not_disturb
      - permissions
      - user_id
      title: User
      type: object
...
    Error:
      description: Error model
      properties:
        error:
          title: error
          type: string
      required:
      - error
      title: Error
      type: object
  securitySchemes:
    Bearer:
      bearerFormat: JWT
      scheme: bearer
      type: http
      x-bearerInfoFunc: account_manager.controllers.security_controller_.info_from_Bearer